+++ src/main/java/com/strangegrotto/wealthdraft/assets/definition/AssetDefinitions.java
@@ -5,6 +5,7 @@ import com.google.common.base.Preconditions;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
 import org.immutables.value.Value;
 
+import java.util.HashSet;
 import java.util.Map;
 
 @WealthdraftImmutableStyle
@@ -12,7 +13,7 @@ import java.util.Map;
 @JsonDeserialize(as = ImmAssetDefinitions.class)
 public interface AssetDefinitions {
     @Value.Parameter
-    Map<String, Asset> getAssets();
+    Map<String, Asset<?, ?>> getAssets();
 
     Map<String, CustomTagDefinition> getCustomTags();
 
@@ -20,39 +21,57 @@ public interface AssetDefinitions {
     default void check() {
         var customTagDefinitions = getCustomTags();
 
-        for (var customTagName : customTagDefinitions.keySet()) {
+        var requiredCustomTags = new HashSet<String>();
+        for (var customTagEntry : customTagDefinitions.entrySet()) {
+            var customTagName = customTagEntry.getKey();
+            var customTagDefinition = customTagEntry.getValue();
+
             Preconditions.checkState(
-                    !IntrinsicAssetTag.getAllNames().contains(customTagName),
+                    !IntrinsicAssetTag.getAllTagNames().contains(customTagName),
                     "Custom tag '%s' collides with an intrinsic tag",
                     customTagName
             );
+            if (customTagDefinition.isRequired()) {
+                requiredCustomTags.add(customTagName);
+            }
         }
 
         var assets = getAssets();
-        for (var assetId : assets.keySet()) {
-            var asset = assets.get(assetId);
-            var customTagsForAsset = asset.getTags();
-            for (var customTag : customTagsForAsset.keySet()) {
-                var customTagValue = customTagsForAsset.get(customTag);
+        for (var assetEntry : assets.entrySet()) {
+            var assetId = assetEntry.getKey();
+            var asset = assetEntry.getValue();
+            var customTagsForAsset = asset.getCustomTags();
+            var requiredTagsNotSeen = new HashSet<>(requiredCustomTags);
+            for (var tagEntry : customTagsForAsset.entrySet()) {
+                var customTagName = tagEntry.getKey();
+                var customTagValue = tagEntry.getValue();
 
                 Preconditions.checkState(
-                        customTagDefinitions.containsKey(customTag),
+                        customTagDefinitions.containsKey(customTagName),
                         "Custom tag '%s' on asset '%s' doesn't match any defined custom tag",
-                        customTag,
+                        customTagName,
                         assetId
                 );
 
-                var customTagDefinition = customTagDefinitions.get(customTag);
+                requiredTagsNotSeen.remove(customTagName);
+
+                var customTagDefinition = customTagDefinitions.get(customTagName);
                 var allowedTagValues = customTagDefinition.getAllowedValues();
                 if (allowedTagValues.size() > 0) {
                     Preconditions.checkState(
                             allowedTagValues.contains(customTagValue),
                             "Tag '%s' for asset '%s' has a value not in the allowed values list for the tag",
-                            customTag,
+                            customTagName,
                             customTagValue
                     );
                 }
             }
+
+            Preconditions.checkState(
+                    requiredTagsNotSeen.size() == 0,
+                    "Asset '%s' doesn't have required tags %s",
+                    assetId,
+                    requiredTagsNotSeen.toString());
         }
     }
 }