+++ src/main/java/com/strangegrotto/wealthdraft/assets/temporal/AssetSnapshot.java
@@ -4,10 +4,15 @@ import com.strangegrotto.wealthdraft.errors.ValOrGerr;
 
 import java.math.BigDecimal;
 
-public interface AssetSnapshot {
+/**
+ * Do not implement this interface directly - instead, inherit from {@link AbstractAssetSnapshot}
+ * @param <CHANGE> The type of change class that this snapshot will consume/produce
+ */
+public interface AssetSnapshot<CHANGE extends AssetChange> {
     BigDecimal getValue();
 
-    AssetSnapshot projectOneMonth();
+    // TODO Make this get an AssetChange instead, and feed into the regular applyChange machinery
+    AssetSnapshot<CHANGE> projectOneMonth();
 
-    ValOrGerr<AssetSnapshot> applyChange(AssetChange change);
+    ValOrGerr<AssetSnapshot<CHANGE>> applyChange(AssetChange change) throws ClassCastException;
 }