+++ src/main/java/com/strangegrotto/wealthdraft/networth/history/AssetsHistoryDeserializer.java
@@ -1,7 +1,6 @@
 package com.strangegrotto.wealthdraft.networth.history;
 
 import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
@@ -18,25 +17,24 @@ import java.util.SortedMap;
 import java.util.TreeMap;
 
 public class AssetsHistoryDeserializer extends JsonDeserializer<AssetsHistory> {
-    private final Map<String, Asset> assets;
+    private final Map<String, Asset<?, ?>> assets;
 
-    public AssetsHistoryDeserializer(Map<String, Asset> assets) {
+    public AssetsHistoryDeserializer(Map<String, Asset<?, ?>> assets) {
         this.assets = assets;
     }
 
     @Override
-    public AssetsHistory deserialize(JsonParser parser, DeserializationContext ctxt) throws IOException, JsonProcessingException {
+    public AssetsHistory deserialize(JsonParser parser, DeserializationContext ctxt) throws IOException {
         // We need to use ObjectMapper.convertValue, but ObjectCodec doesn't have it on it
         var mapper = (ObjectMapper) parser.getCodec();
         Map<String, Map<LocalDate, Map<String, String>>> raw = parser.readValueAs(
                 new TypeReference<Map<String, Map<LocalDate, Map<String, String>>>>(){}
         );
 
-        LocalDate today = LocalDate.now();
-
-        var parsedAssetSnapshots = new HashMap<String, SortedMap<LocalDate, AssetSnapshot>>();
-        for (var assetId : raw.keySet()) {
-            var unparsedSnapshotsForAsset = raw.get(assetId);
+        var parsedAssetSnapshots = new HashMap<String, SortedMap<LocalDate, AssetSnapshot<?>>>();
+        for (var rawEntry : raw.entrySet()) {
+            var assetId = rawEntry.getKey();
+            var unparsedSnapshotsForAsset = rawEntry.getValue();
             Preconditions.checkState(
                     this.assets.containsKey(assetId),
                     "Asset ID '%s' doesn't match any known asset"
@@ -44,9 +42,10 @@ public class AssetsHistoryDeserializer extends JsonDeserializer<AssetsHistory> {
             var asset = this.assets.get(assetId);
             var snapshotType = asset.getSnapshotType();
 
-            var parsedSnapshotsForAsset = new TreeMap<LocalDate, AssetSnapshot>();
-            for (var date : unparsedSnapshotsForAsset.keySet()) {
-                var unparsedSnapshot = unparsedSnapshotsForAsset.get(date);
+            var parsedSnapshotsForAsset = new TreeMap<LocalDate, AssetSnapshot<?>>();
+            for (var unparsedSnapshotEntry : unparsedSnapshotsForAsset.entrySet()) {
+                var date = unparsedSnapshotEntry.getKey();
+                var unparsedSnapshot = unparsedSnapshotEntry.getValue();
                 var parsedSnapshot = mapper.convertValue(unparsedSnapshot, snapshotType);
                 parsedSnapshotsForAsset.put(date, parsedSnapshot);
             }
@@ -54,5 +53,5 @@ public class AssetsHistoryDeserializer extends JsonDeserializer<AssetsHistory> {
         }
 
         return ImmAssetsHistory.of(this.assets, parsedAssetSnapshots);
-    };
+    }
 }