+++ src/test/java/com/strangegrotto/wealthdraft/assets/definition/IntrinsicAssetTagTest.java
@@ -1,16 +1,11 @@
 package com.strangegrotto.wealthdraft.assets.definition;
 
 import com.strangegrotto.wealthdraft.EnumDistinctnessVerifier;
-import com.strangegrotto.wealthdraft.assetimpls.AssetTypeTagValue;
-import org.junit.Assert;
 import org.junit.Test;
-
-import java.util.Arrays;
-import java.util.stream.Collectors;
 
 public class IntrinsicAssetTagTest {
     @Test
     public void ensureNoDuplicates() {
-        EnumDistinctnessVerifier.verifyDistinct(IntrinsicAssetTag.values(), IntrinsicAssetTag::getName);
+        EnumDistinctnessVerifier.verifyDistinct(IntrinsicAssetTag.values(), IntrinsicAssetTag::getTagName);
     }
 }